+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## [1.3.0] - 2023-04-11
+
+### Added
+
+- Creation Login component.
+
+## [1.2.0] - 2023-04-10
+
+### Added
+
+- Creation Registration component.
+
+## [1.1.0] - 2023-04-09
+
+### Added
+
+- Creation footer component.
+
 ## [1.0.0] - 2023-04-08
 
 ### Added